refactor: add handleCommandExecution function
