--- src/commands/commands.constants.ts
@@ -1,7 +1,0 @@
-export enum Commands {
-  Summary = '/summary',
-  Consensus = '/consensus',
-  Disagreements = '/disagreements',
-  Compromises = '/compromises',
-  DraftProposal = '/proposal',
-}
